Refactor common functionality between players
Refactor the actual lottie-web interfacing code into a common,
framework-agnostic `LottiePlayer` class. `LottiePlayer` mounts a
`lottie-web` instance to a given container element and provides a
uniform interface to it.

Modify the Web and React components to take advantage of this new class
to avoid logic duplication. This allows the React component to provide
access to this player directly, via a ref, and readily expose some more
parameters.

This also includes a fix for an issue in the web component where both
`connectedCallback` and `attributeChanged` would trigger on init,
causing a race condition.

(The above is fixed by not running both, and also by ensuring that
LottiePlayer JSON file fetches "cancel" each other, to avoid an
intermediate state where the fetch is in progress, the player has not
yet been 'initialized', and another initialization takes over).

Bonus: use XMLHttpRequest instead of fetch() to ensure maximum compat
between browsers.