fix: Better blend + transition interaction
Currently, transitioning into or out of a state with morphing
(blending) enabled doesn't work well, since we ignore the morphing
configuration of the state when interpolating the transition, causing
abrupt changes.

On the other hand, our morphing representation is not flexible enough to
support a "proper" solution (see comments).

Workaround the most common case, which will still leave some edge-cases,
and make a note to fix this in a more permanent way (with a larger
refactoring of how morphing works) sometime in the future.

Also, clean up some redundant renderFrame() calls.